Enable v1.5.0 feature in gke cluster provisioner
